+++ timesheet/package.json
@@ -23,9 +23,10 @@
     "@fullcalendar/angular": "^6.1.10",
     "@fullcalendar/core": "^6.1.10",
     "@fullcalendar/daygrid": "^6.1.10",
+    "bootstrap": "^5.3.2",
     "bootstrap-icons": "^1.11.3",
     "firebase": "^9.0.0",
-    "ngx-toastr": "^12.0.0",
+    "ngx-toastr": "^18.0.0",
     "primeng": "^17.3.1",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
