+++ timesheet/src/app/component/attendance/attendance/attendance.component.html
@@ -1,1 +1,3 @@
-<p>attendance works!</p>
+<button class="btn btn-danger" (click)="toggleWeekends()">toggle weekends</button>
+
+<full-calendar [options]="calendarOptions"></full-calendar>
