+++ timesheet/src/app/component/attendance/attendance/attendance.module.ts
@@ -1,9 +1,11 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { AttendanceRoutingModule } from './attendance-routing.module';
+import { AttendanceComponent } from './attendance.component';
+import { FullCalendarModule } from '@fullcalendar/angular';
 
 @NgModule({
-  declarations: [],
-  imports: [CommonModule, AttendanceRoutingModule],
+  declarations: [AttendanceComponent],
+  imports: [CommonModule, AttendanceRoutingModule, FullCalendarModule],
 })
 export class AttendanceModule {}
