--- timesheet/src/app/component/time-sheetmanegement/timesheet/timesheet.component.html
@@ -1,10 +1,0 @@
-
-
-
-<button class="btn btn-danger" (click)="toggleWeekends()">toggle weekends</button>
-
-<full-calendar [options]="calendarOptions"></full-calendar>
-
-
-
-
